+++ packages/codegen.go/src/core/operations.ts
@@ -996,7 +996,13 @@ function createProtocolRequest(azureARM: boolean, method: go.MethodType | go.Nex
     for (const pp of methodParamGroups.pathParams) {
       let paramValue: string;
       let optionalPathSep = false;
-      if (pp.style !== 'optional') {
+      if (pp.style === 'literal') {
+        // literals are always scalar types and require no empty checks
+        paramValue = helpers.formatParamValue(pp, imports, indent);
+      } else if (pp.style === 'required' || pp.location === 'client') {
+        // NOTE: we include client params here since they behave
+        // like required params (i.e. not grouped).
+
         // emit check to ensure path param isn't an empty string
         if (pp.kind === 'pathScalarParam') {
           const choiceIsString = function (type: go.PathScalarParameterType): boolean {
@@ -1027,6 +1033,13 @@ function createProtocolRequest(azureARM: boolean, method: go.MethodType | go.Nex
           text += `${indent.pop().get()}}\n`;
           paramValue = joinedParamName;
         }
+      } else if (go.isClientSideDefault(pp.style)) {
+        const defaultValue = naming.uncapitalize(pp.name) + 'Default';
+        text += `${indent.get()}${defaultValue} := ${helpers.formatLiteralValue(pp.style.defaultValue, true)}\n`;
+        text += emitParamGroupCheck(pp);
+        text += `${indent.push().get()}${defaultValue} = ${helpers.getParamName(pp)}\n`;
+        text += `${indent.pop().get()}}\n`;
+        paramValue = helpers.formatValue(defaultValue, pp.type, imports);
       } else {
         // param isn't required, so emit a local var with
         // the correct default value, then populate it with